refactor(desktop): Extract logs and crash dump logic from commands.rs
