+++ apps/desktop/src-tauri/src/logs/mod.rs
@@ -1,0 +1,5 @@
+pub mod crash_dumps;
+pub mod log_manager;
+
+pub use crash_dumps::CrashDumpInfo;
+pub use log_manager::LogInfo;
